add a  branch for not moving the cursor

+++ truck-modeling/examples/framework.rs
@@ -99,6 +99,9 @@ impl ShapeViewer {
             let position = Vector2::new(position.x, position.y);
             if let Some(ref prev_position) = self.prev_cursor {
                 let dir2d = &position - prev_position;
+                if dir2d.so_small() {
+                    return Self::default_control_flow();
+                }
                 let mut axis = dir2d[1] * matrix[0].truncate();
                 axis += dir2d[0] * &matrix[1].truncate();
                 axis /= axis.magnitude();
